Refactoring Post2dWindowNodeVectorParticleGroupUnstructuredDataItem::SetSettingCommand moved to separate files.
